Split binds.rs/heap_result_arm.rs/tail.rs/tests_part1-2.rs to clear max-lines
