+++ crates/almide-mir/src/lower/tests.rs
@@ -6,4 +6,6 @@
     use crate::{verify_ownership, ViolationKind};
     use almide_lang::types::constructor::TypeConstructorId;
     include!("tests_part1.rs");
+    include!("tests_part1_b.rs");
     include!("tests_part2.rs");
+    include!("tests_part2_b.rs");
